Refactoring the main calc function
Changing the logic of the calculations, instead of using eval, I use conditional calculations upon operator. Mainly because of the negative numbers handling and multiple = operator clicks.

In operators component, added a condition to check if the last item in array is a number or operator.

boilercode for save button with styling