Storage component, delete item refactoring
Splited the components to new files in storage folder accordingly.

Handle Delete function takes care now in renaming the keys and name property of the object. The previous version did delete the right key object, but with new calculation and item creation, it would missname the new object and cause it to rewrite othe objects.