Item header renaming function
Logic to renaming headers of the calc storage items. each item has its own state for the header which is easier to update than directly the storage object itself on onchange.

Changed the header to be an input field instead of div, because of the problem with content editable and the state handling from react.