Storage item comment, number renaming
Functions to support comments renaming and proper object/ array handling. Changed the text types to input elements for better value handling.

The changing of numbers in the item container has to have two functions, one for updating the objects array values and the second for recalculating the sum value of the object.

Changing the styling of the storage window to be in grid mode