Sum item copy, Storage, item component refactoring
Sum item in the calc storage - clicking on the sum will now copy the sum to the display object so the user can combine the containers between each other.

Refacatoring the storage window, mainly how the items are being rendered. From having a one flexbox window to three separate columns to have the masonry style.

Refactored how the items are rendered using input instead of paragraphs only to be able to better handle the values and their updates.

Created extra function to split the passed object to attach them to the correct column.