window Width, media queries, bug fixing
The tailwind media queries didn't work properly so I added functions to get the page width on each render so if the user is playing with resizing the window the right icons appear and css is applied.

New and better conditions for the save and menu icons in the display component.

Bug fixing - the new conditions prevent the user to be able to save the calculations when last item in the display.mem array is not a number or the array is empty